Cleaned up Reference.h

+++ Silver/Source/Silver/Renderer/Pipeline.cpp
@@ -61,8 +61,8 @@ namespace Silver {
 		VkViewport viewport{};
 		viewport.x = 0.0f;
 		viewport.y = 0.0f;
-		viewport.width = extent.width;
-		viewport.height = extent.height;
+		viewport.width = (float)extent.width;
+		viewport.height = (float)extent.height;
 		viewport.minDepth = 0.0f;
 		viewport.maxDepth = 1.0f;
